backend/requirements.txt
tensorflow==2.10.0
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.23.5
Pillow==10.0.0
scikit-learn==1.3.2
h5py